backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
python-dotenv
pydantic-settings
asyncpg
bcrypt==4.0.1
pdfplumber
google-generativeai
